// src/generated/PostInfoList.ts
export interface PostInfoList {
  forumId: string;
  threadId: string;
  postId: string;
  createTime: number;
  forumName: string;
  title: string;
  content: PostInfoList_PostInfoContent[];
  userName: string;
  media: Media[];
  replyNum: number;
  userId: string;
  userPortrait: string;
  voiceInfo: Voice[];
  threadType: string;
  pollInfo: PollInfo | undefined;
  videoInfo: VideoInfo | undefined;
  freqNum: number;
  nameShow: string;
  shareNum: number;
  agree: Agree | undefined;
  isShareThread: number;
  firstPostContent: PbContent[];
}

export interface PostInfoList_PostInfoContent {
  postContent: PostInfoList_PostInfoContent_Abstract[];
  createTime: string;
  postType: string;
  postId: string;
}

export interface PostInfoList_PostInfoContent_Abstract {
  type: number;
  text: string;
  link: string;
  duringTime: string;
  voiceMd5: string;
}
// ...
export const PostInfoList_PostInfoContent: MessageFns<PostInfoList_PostInfoContent> = {
  encode(message: PostInfoList_PostInfoContent, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.postContent) {
      PostInfoList_PostInfoContent_Abstract.encode(v!, writer.uint32(10).fork()).join();
    }
    if (message.createTime !== "0") {
      writer.uint32(16).uint64(message.createTime);
    }
    if (message.postType !== "0") {
      writer.uint32(24).uint64(message.postType);
    }
    if (message.postId !== "0") {
      writer.uint32(32).uint64(message.postId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PostInfoList_PostInfoContent {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePostInfoList_PostInfoContent();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.postContent.push(PostInfoList_PostInfoContent_Abstract.decode(reader, reader.uint32()));
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.createTime = reader.uint64().toString();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.postType = reader.uint64().toString();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.postId = reader.uint64().toString();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PostInfoList_PostInfoContent {
    return {
      postContent: globalThis.Array.isArray(object?.postContent)
        ? object.postContent.map((e: any) => PostInfoList_PostInfoContent_Abstract.fromJSON(e))
        : globalThis.Array.isArray(object?.post_content)
        ? object.post_content.map((e: any) => PostInfoList_PostInfoContent_Abstract.fromJSON(e))
        : [],
      createTime: isSet(object.createTime)
        ? globalThis.String(object.createTime)
        : isSet(object.create_time)
        ? globalThis.String(object.create_time)
        : "0",
      postType: isSet(object.postType)
        ? globalThis.String(object.postType)
        : isSet(object.post_type)
        ? globalThis.String(object.post_type)
        : "0",
      postId: isSet(object.postId)
        ? globalThis.String(object.postId)
        : isSet(object.post_id)
        ? globalThis.String(object.post_id)
        : "0",
    };
  },

  toJSON(message: PostInfoList_PostInfoContent): unknown {
    const obj: any = {};
    if (message.postContent?.length) {
      obj.postContent = message.postContent.map((e) => PostInfoList_PostInfoContent_Abstract.toJSON(e));
    }
    if (message.createTime !== "0") {
      obj.createTime = message.createTime;
    }
    if (message.postType !== "0") {
      obj.postType = message.postType;
    }
    if (message.postId !== "0") {
      obj.postId = message.postId;
    }
    return obj;
  },

  create(base?: DeepPartial<PostInfoList_PostInfoContent>): PostInfoList_PostInfoContent {
    return PostInfoList_PostInfoContent.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PostInfoList_PostInfoContent>): PostInfoList_PostInfoContent {
    const message = createBasePostInfoList_PostInfoContent();
    message.postContent = object.postContent?.map((e) => PostInfoList_PostInfoContent_Abstract.fromPartial(e)) || [];
    message.createTime = object.createTime ?? "0";
    message.postType = object.postType ?? "0";
    message.postId = object.postId ?? "0";
    return message;
  },
};

function createBasePostInfoList_PostInfoContent_Abstract(): PostInfoList_PostInfoContent_Abstract {
  return { type: 0, text: "", link: "", duringTime: "", voiceMd5: "" };
}

export const PostInfoList_PostInfoContent_Abstract: MessageFns<PostInfoList_PostInfoContent_Abstract> = {
  encode(message: PostInfoList_PostInfoContent_Abstract, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.text !== "") {
      writer.uint32(18).string(message.text);
    }
    if (message.link !== "") {
      writer.uint32(26).string(message.link);
    }
    if (message.duringTime !== "") {
      writer.uint32(50).string(message.duringTime);
    }
    if (message.voiceMd5 !== "") {
      writer.uint32(58).string(message.voiceMd5);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): PostInfoList_PostInfoContent_Abstract {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBasePostInfoList_PostInfoContent_Abstract();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.text = reader.string();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.link = reader.string();
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.duringTime = reader.string();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.voiceMd5 = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): PostInfoList_PostInfoContent_Abstract {
    return {
      type: isSet(object.type) ? globalThis.Number(object.type) : 0,
      text: isSet(object.text) ? globalThis.String(object.text) : "",
      link: isSet(object.link) ? globalThis.String(object.link) : "",
      duringTime: isSet(object.duringTime)
        ? globalThis.String(object.duringTime)
        : isSet(object.during_time)
        ? globalThis.String(object.during_time)
        : "",
      voiceMd5: isSet(object.voiceMd5)
        ? globalThis.String(object.voiceMd5)
        : isSet(object.voice_md5)
        ? globalThis.String(object.voice_md5)
        : "",
    };
  },

  toJSON(message: PostInfoList_PostInfoContent_Abstract): unknown {
    const obj: any = {};
    if (message.type !== 0) {
      obj.type = Math.round(message.type);
    }
    if (message.text !== "") {
      obj.text = message.text;
    }
    if (message.link !== "") {
      obj.link = message.link;
    }
    if (message.duringTime !== "") {
      obj.duringTime = message.duringTime;
    }
    if (message.voiceMd5 !== "") {
      obj.voiceMd5 = message.voiceMd5;
    }
    return obj;
  },

  create(base?: DeepPartial<PostInfoList_PostInfoContent_Abstract>): PostInfoList_PostInfoContent_Abstract {
    return PostInfoList_PostInfoContent_Abstract.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<PostInfoList_PostInfoContent_Abstract>): PostInfoList_PostInfoContent_Abstract {
    const message = createBasePostInfoList_PostInfoContent_Abstract();
    message.type = object.type ?? 0;
    message.text = object.text ?? "";
    message.link = object.link ?? "";
    message.duringTime = object.duringTime ?? "";
    message.voiceMd5 = object.voiceMd5 ?? "";
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}